Add a method that sends a command to each of the cluster nodes
As Redis does not allow cross-slot commands, there are situations when
it becomes necessary to send a single command to all nodes.

This project already supports cross-slot execution of the command
_MGET_ and it would be natural to support other non-cross-slot commands
such as _SCAN_.

The file **each_test.go** contains an example that sends a _SCAN_ command
to all nodes of the cluster. Also note that **each_test.go** is
formatted for `godoc`, so that it becames an example embedded in the
documentation.

Please review and let me know if there are any issues.

Thank you.